requirements.txt
django>=5.0
mysqlclient>=2.2.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
anthropic>=0.25.0
mcp>=1.0.0
python-dotenv>=1.0.0